Moved logic for FireBow & Hiding
