+++ src/main/java/net/laboulangerie/laboulangeriemmo/core/abilities/thehunter/hiding/InvisiblePlayer.java
@@ -1,4 +1,4 @@
-package net.laboulangerie.laboulangeriemmo.core.thehunter.hiding;
+package net.laboulangerie.laboulangeriemmo.core.abilities.thehunter.hiding;
 
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Entity;
